back/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
supabase==2.3.0
paho-mqtt>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
